pyproject.toml
[tool.poetry]
name = "wuyubot"
version = "0.1.0"
description = "一个普通的nonebot实现"
authors = ["yeqiu6080 <yeqiu6080@outlook.com>"]
license = "AGPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot-adapter-onebot = "^2.4.6"
nonebot2 = {extras = ["fastapi","httpx","websockets"], version = "^2.4.1"}
noneprompt = "^0.1.9"
pyyaml = "^6.0.2"
redis = "^5.2.1"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["wuyubot/plugins"]
builtin_plugins = []

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
